# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /materials/{batchId}/publish-to-market:
    post:
      operationId: MaterialsMarketController_publishToMarket
      parameters:
        - name: batchId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToMarketDto"
      responses:
        "201":
          description: ""
      tags: &a12
        - MaterialsMarket
  /materials/market/{listingId}:
    patch:
      operationId: MaterialsMarketController_updateMarketListing
      parameters:
        - name: listingId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMarketListingDto"
      responses:
        "200":
          description: ""
      tags: *a12
  /materials/market/{listingId}/withdraw:
    post:
      operationId: MaterialsMarketController_withdrawMarketListing
      parameters:
        - name: listingId
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags: *a12
...
info:
  title: PACUL Backend API
  description: REST API for the PACUL circular waste marketplace (hackathon MVP).
  version: 1.0.0-hackathon
  contact: {}
...
servers:
  - url: http://localhost:4000
    description: Local development
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Supabase access token from sign-in
  schemas:
...
    PublishToMarketDto:
      type: object
      properties: {}
    UpdateMarketListingDto:
      type: object
      properties: {}
...